Sending word to Anki works 	but only one translate available instead one OR two
