+++ package.json
@@ -13,6 +13,7 @@
         "@react-native-community/async-storage": "^1.8.1",
         "axios": "^0.19.2",
         "babel-preset-react-native": "^4.0.1",
+        "jsonfn": "^0.31.0",
         "native-base": "^2.13.8",
         "node-fetch": "^2.6.0",
         "react": "16.13.0",
