+++ src/actions/api/dictionary.js
@@ -13,6 +13,7 @@ import {
 } from './urban-dictionary';
 import {createDictionary} from '../dictionary/create-dictionary';
 import {setFields} from '../anki-set-actions';
+import {sendWord, submit} from '../form-actions';
 
 const getAvailableApi = (apiArray = []) => {
   for (const api of apiArray) {
@@ -48,7 +49,7 @@ export const wordInfo = word => async dispatch => {
     await dispatch(setAvailableApi(api1));
     const wordDictionary = await createDictionary(api1);
     dispatch(setDictioanry(wordDictionary));
-    dispatch(setFields(wordDictionary));
+    sendWord(setFields(wordDictionary));
   } catch (e) {
     console.log(e);
   }
