+++ src/actions/filesystem.js
@@ -3,10 +3,17 @@ import {
   ANKILAN_NOTE_CREATOR,
   ANKILAN_NOTE_TEMPLATE,
 } from '../constants/anki-constants';
+import JSONfn from 'jsonfn';
+
 const sendDataToLocaleStorage = async (creator, template) => {
   try {
-    await AsyncStorage.setItem(ANKILAN_NOTE_CREATOR, JSON.stringify(creator));
-    await AsyncStorage.setItem(ANKILAN_NOTE_TEMPLATE, JSON.stringify(template));
+    await AsyncStorage.clear();
+
+    await AsyncStorage.setItem(ANKILAN_NOTE_CREATOR, JSONfn.stringify(creator));
+    await AsyncStorage.setItem(
+      ANKILAN_NOTE_TEMPLATE,
+      JSONfn.stringify(template),
+    );
   } catch (e) {
     // saving error
     alert('Error while syncing with filesystem');
@@ -18,9 +25,10 @@ export default sendDataToLocaleStorage;
 export const getTemplate = async () => {
   try {
     const value = await AsyncStorage.getItem(ANKILAN_NOTE_TEMPLATE);
+    console.log(value);
     if (value !== null) {
       // value previously stored
-      return value;
+      return JSONfn.parse(value);
     }
   } catch (e) {
     // error reading value
@@ -29,10 +37,10 @@ export const getTemplate = async () => {
 
 export const getCreator = async () => {
   try {
-    const value = await AsyncStorage.getItem(ANKILAN_NOTE_TEMPLATE);
+    const value = await AsyncStorage.getItem(ANKILAN_NOTE_CREATOR);
     if (value !== null) {
       // value previously stored
-      return value;
+      return JSONfn.parse(value);
     }
   } catch (e) {
     // error reading value
