+++ src/reducers/anki-reducer.js
@@ -20,7 +20,7 @@ const initialState = {
   mainFieldIsAvailable: false,
   fieldList: [],
   ankiLanModelIsAlreadyExists: false,
-  ankiLanModelName: 'AnkiLan_final1',
+  ankiLanModelName: '1AnkiLan1111',
   noteCreator: {},
   noteTemplate: [],
 };
